合并通用库文件
