+++ WechatJump/Config.py
@@ -1,0 +1,32 @@
+# coding: utf-8
+import os
+import sys
+import json
+import re
+
+def open_accordant_config():
+    screen_size = _get_screen_size()
+    config_file = "{path}/config/{screen_size}/config.json".format(
+        path=sys.path[0],
+        screen_size=screen_size
+    )
+    if os.path.exists(config_file):
+        with open(config_file, 'r') as f:
+            print("Load config file from {}".format(config_file))
+            return json.load(f)
+    else:
+        with open('{}/config/default.json'.format(sys.path[0]), 'r') as f:
+            print("Load default config")
+            return json.load(f)
+
+
+def _get_screen_size():
+    size_str = os.popen('adb shell wm size').read()
+    if not size_str:
+        print('请安装ADB及驱动并配置环境变量')
+        sys.exit()
+    m = re.search('(\d+)x(\d+)', size_str)
+    if m:
+        width = m.group(1)
+        height = m.group(2)
+        return "{height}x{width}".format(height=height, width=width)
